Refactoring scrapers/classes/processors
