+++ .gitignore
@@ -1,5 +1,5 @@
 node_modules/
-cache*
+cache_
 dist
 public
 *.log
