+++ scrapers/classes/main.js
@@ -26,7 +26,7 @@ class Main {
   // Runs the processors over a termDump.
   // The input of this function should be the output of restructureData, above.
   // The updater.js calls into this function to run the processors over the data scraped as part of the processors.
-  runProcessors(dump) {
+  runProcessors(dump) { //ParsedTermSR
     // Run the processors, sequentially
     markMissingPrereqs.go(dump);
     termStartEndDate.go(dump);
